metadata-api/requirements.txt
aiohttp==3.10.11
cachetools==5.3.2
fastapi==0.104.1
lxml==4.9.3
pillow==10.3.0
uvicorn==0.24.0